build: Updating the build process to use makefile
Based on commit 27b71778931c3f107eaaf5feacaaa5fa61d9c10f, the idea of
this commit is to transform the whole build chain to use makefile
instead of the actual hacky generate-dev-env.sh + custom scripts.

It have the benefits of improving parallelism, building in sub-directories and simple make commands.

"base" is now moved into the root (/) of the project as it's no more specific to a custom ceph release.

<TBU>